Fix problem with creation of project
